fix(zombie_processes): suggest waiting on a discarded Result<Child>
Result has no wait method, so the suggestion replaces the statement with an
if let Ok block instead of appending like the Child case.
