+++ tests/ui/zombie_processes.stderr
@@ -127,25 +127,5 @@ LL |             x.wait().unwrap();
    = note: not doing so might leave behind zombie processes
    = note: see https://doc.rust-lang.org/stable/std/process/struct.Child.html#warning
 
-error: spawned process is never `wait()`ed on
-  --> tests/ui/zombie_processes.rs:217:17
-   |
-LL |         let _ = Command::new("true").spawn();
-   |                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-   |
-   = help: consider calling `.wait()`
-   = note: not doing so might leave behind zombie processes
-   = note: see https://doc.rust-lang.org/stable/std/process/struct.Child.html#warning
-
-error: spawned process is never `wait()`ed on
-  --> tests/ui/zombie_processes.rs:221:9
-   |
-LL |         Command::new("true").spawn();
-   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-   |
-   = help: consider calling `.wait()`
-   = note: not doing so might leave behind zombie processes
-   = note: see https://doc.rust-lang.org/stable/std/process/struct.Child.html#warning
-
-error: aborting due to 10 previous errors
+error: aborting due to 8 previous errors
